Update seamly2d.pro to fix xpdf dependency
this is a quick and dirty test to see if the resulting build will load on Ubuntu focal.  Currently both Valentina and Seamly2d fail to install on that release due to xpdf no longer being loaded as an exposed symbol in the default OS package.  This may be OBE after April 23.  If it is still an issue once the official release of 20.04 lts is finalized, then a real fix will need to be created.  This particular patch would potentially break the MAC version, so more research would be needed