remove xpdf dependency
Update dist/debian/control to remove xpdf dependency.  This primarily is an issue in the build performed on launchpad and affects users of FOCAL (Ubuntu 20.04 LTS).  This will probably not show up as an issue once appimage version is created, nor will it affect users of releases of Ubuntu earlier than 18.10