Change "FlippingBy" to MirrorBy - issue#345
This refactors the code to change "flippimgBy" to "MirrorBy" for the FlippingByLine and FlippingByAxis tools. It also freshens up the dialogs in the Operations group of tools, as well as fixing some inconsistancies with those tools  in the Properties Editor.

Also adds a clear button for the formula field in the Edit Formula dialog.