Add color & visibility to grainlines and piece labels
Update the grainline & labels icons

This adds color to grainlines and pattern piece labels, as well as adding the ability to globally show / hide all grainlines and / or pattern piece labels.

This also fixes the issue of correctly displaying "text as paths" in the piece labels to display just the path (outline) of the text with no brush fill.

Add line weight to grainlines