build: integrate all ci flows into one yml to reuse build steps
To avoid copy-paste errors between build-auto-prerelease-on-merge.yml, build-release.yml
and build-auto-qttest-on-pullrequest.yml, merge them into one ci.yml and guard
the release creation with "if:" based on the github trigger.

At the same time replace the unmaintained actions/create-release and actions/upload-release-asset
with softprops/action-gh-release. This also comes with automatic changelog for the (pre-)releases
based on the pull-request titles.

In addition, as the version is now created once and then reused via output and env, all builds and
the tag have exactly the same version string.