split publishing into two paths
Publishing has been split into two tasks in CI:

One for the case of signed executables,
and one for the case of unsigned executables.

The reason for this is that there's a conditional
part of the workflow, which allows bypassing signing
if needed. But publishing signed executables has a dependendcy
on the signing step, which is not needed for unsigned executables
since the unsigned executables are released after the build step.

Conditional dependencies are not supported in GitHub Actions,
so the publishing of signed executables has been moved
to a separate task.