bug: fix variables table refresh (#1408)
This fixes several cases where the Variables Table dialog does not refresh when tools are deleted, or as the case is close when the pattern is closed.

Adds emitting FullUpdateFromFile signals to Undo's for when either a tool is deleted, or Undo'ing a tool that was added... as well as Redo'ing a when a tool is undone. This signal is then handled by the DialogVariables to refresh the data in the tables.

Added new VAbstractPattern::patternClosed() signal which is emmited when a pattern is closed, which is then handled by DialogVariables to close the dialog instead of leaving it open filled with invalid tables.

Signed-off-by: DSCaskey <31944718+DSCaskey@users.noreply.github.com>

---------

Signed-off-by: DSCaskey <31944718+DSCaskey@users.noreply.github.com>